# packages/api/generated/.docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Vault Hub Server
  version: 1.0.6
servers:
- url: /
paths:
...
  /api/vaults:
    get:
      description: Get all vaults for the current user
      operationId: getVaults
      parameters:
      - description: "Number of vaults per page (default 20, max 1000)"
        explode: true
        in: query
        name: pageSize
        required: false
        schema:
          default: 20
          maximum: 1000
          minimum: 1
          type: integer
        style: form
      - description: "Page index, starting from 1 (default 1)"
        explode: true
        in: query
        name: pageIndex
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultsResponse"
          description: Paginated list of vaults
      tags:
      - Vault
    post:
      description: Create a new vault
      operationId: createVault
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVaultRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"
          description: Vault created successfully
      tags:
      - Vault
  /api/vaults/{uniqueId}:
    delete:
      description: Delete a vault
      operationId: deleteVault
      parameters:
      - description: Vault Unique ID
        explode: false
        in: path
        name: uniqueId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Vault deleted successfully
      tags:
      - Vault
    get:
      description: Get a specific vault by Unique ID
      operationId: getVault
      parameters:
      - description: Vault Unique ID
        explode: false
        in: path
        name: uniqueId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"
          description: Vault details
      tags:
      - Vault
    put:
      description: Update a vault
      operationId: updateVault
      parameters:
      - description: Vault Unique ID
        explode: false
        in: path
        name: uniqueId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVaultRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"
          description: Vault updated successfully
      tags:
      - Vault
  /api/audit-logs:
    get:
      description: Get audit logs with optional filtering and pagination
      operationId: getAuditLogs
      parameters:
      - description: Filter logs from this date (ISO 8601 format)
        explode: true
        in: query
        name: startDate
        required: false
        schema:
          format: date-time
          type: string
        style: form
      - description: Filter logs until this date (ISO 8601 format)
        explode: true
        in: query
        name: endDate
        required: false
        schema:
          format: date-time
          type: string
        style: form
      - description: Filter logs by vault unique ID
        explode: true
        in: query
        name: vaultUniqueId
        required: false
        schema:
          type: string
        style: form
      - description: "Number of logs per page (default 100, max 1000)"
        explode: true
        in: query
        name: pageSize
        required: true
        schema:
          default: 20
          maximum: 1000
          minimum: 1
          type: integer
        style: form
      - description: "Page index, starting from 0 (default 0)"
        explode: true
        in: query
        name: pageIndex
        required: true
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogsResponse"
          description: List of audit logs
      tags:
      - Audit
...
  /api/cli/vaults:
    get:
      description: Get all vaults accessible by API key
      operationId: getVaultsByAPIKey
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/VaultLite"
                type: array
          description: List of vaults accessible by API key
      security:
      - ApiKeyAuth: []
      tags:
      - Cli
  /api/cli/vault/{uniqueId}:
    get:
      description: Get a specific vault by Unique ID using API key
      operationId: getVaultByAPIKey
      parameters:
      - description: Vault Unique ID
        explode: false
        in: path
        name: uniqueId
        required: true
        schema:
          type: string
        style: simple
      - description: Enable client-side encryption (server returns encrypted value)
        explode: false
        in: header
        name: X-Enable-Client-Encryption
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"
          description: Vault details
      security:
      - ApiKeyAuth: []
      tags:
      - Cli
  /api/cli/vault/name/{name}:
    get:
      description: Get a specific vault by name using API key
      operationId: getVaultByNameAPIKey
      parameters:
      - description: Vault name
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      - description: Enable client-side encryption (server returns encrypted value)
        explode: false
        in: header
        name: X-Enable-Client-Encryption
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"
          description: Vault details
      security:
      - ApiKeyAuth: []
      tags:
      - Cli
components:
  schemas:
...
    VaultLite:
      example:
        name: name
        description: description
        category: category
        uniqueId: uniqueId
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        uniqueId:
          description: Unique identifier for the vault
          type: string
        name:
          description: Human-readable name
          type: string
        description:
          description: Human-readable description
          type: string
        category:
          description: Category/type of vault
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - name
      - uniqueId
      type: object
    Vault:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        name: name
        description: description
        category: category
        userId: 0
        value: value
        uniqueId: uniqueId
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        uniqueId:
          description: Unique identifier for the vault
          type: string
        userId:
          description: ID of the user who owns this vault
          format: int64
          type: integer
        name:
          description: Human-readable name
          type: string
        value:
          description: Encrypted value
          type: string
        description:
          description: Human-readable description
          type: string
        category:
          description: Category/type of vault
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - name
      - uniqueId
      - value
      type: object
    CreateVaultRequest:
      example:
        name: name
        description: description
        category: category
        value: value
      properties:
        name:
          description: Human-readable name
          maxLength: 255
          minLength: 1
          type: string
        value:
          description: Value to be encrypted and stored
          minLength: 1
          type: string
        description:
          description: Human-readable description
          maxLength: 500
          type: string
        category:
          description: Category/type of vault
          maxLength: 100
          type: string
      required:
      - name
      - value
      type: object
    UpdateVaultRequest:
      example:
        name: name
        description: description
        category: category
        value: value
      properties:
        name:
          description: Human-readable name
          maxLength: 255
          minLength: 1
          type: string
        value:
          description: Value to be encrypted and stored
          minLength: 1
          type: string
        description:
          description: Human-readable description
          maxLength: 500
          type: string
        category:
          description: Category/type of vault
          maxLength: 100
          type: string
      type: object
    AuditLogsResponse:
      example:
        auditLogs:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          apiKey:
            createdAt: 2000-01-23T04:56:07.000+00:00
            lastUsedAt: 2000-01-23T04:56:07.000+00:00
            name: name
            vaults:
            - name: name
              description: description
              category: category
              uniqueId: uniqueId
              updatedAt: 2000-01-23T04:56:07.000+00:00
            - name: name
              description: description
              category: category
              uniqueId: uniqueId
              updatedAt: 2000-01-23T04:56:07.000+00:00
            id: 0
            isActive: true
            expiresAt: 2000-01-23T04:56:07.000+00:00
            updatedAt: 2000-01-23T04:56:07.000+00:00
          ipAddress: ipAddress
          action: read_vault
          userAgent: userAgent
          source: web
          vault:
            name: name
            description: description
            category: category
            uniqueId: uniqueId
            updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          apiKey:
            createdAt: 2000-01-23T04:56:07.000+00:00
            lastUsedAt: 2000-01-23T04:56:07.000+00:00
            name: name
            vaults:
            - name: name
              description: description
              category: category
              uniqueId: uniqueId
              updatedAt: 2000-01-23T04:56:07.000+00:00
            - name: name
              description: description
              category: category
              uniqueId: uniqueId
              updatedAt: 2000-01-23T04:56:07.000+00:00
            id: 0
            isActive: true
            expiresAt: 2000-01-23T04:56:07.000+00:00
            updatedAt: 2000-01-23T04:56:07.000+00:00
          ipAddress: ipAddress
          action: read_vault
          userAgent: userAgent
          source: web
          vault:
            name: name
            description: description
            category: category
            uniqueId: uniqueId
            updatedAt: 2000-01-23T04:56:07.000+00:00
        pageIndex: 5
        pageSize: 1
        totalCount: 6
      properties:
        auditLogs:
          items:
            $ref: "#/components/schemas/AuditLog"
          type: array
        totalCount:
          description: Total number of logs matching the filter criteria
          type: integer
        pageSize:
          description: Number of logs per page
          type: integer
        pageIndex:
          description: Current page index (starting from 0)
          type: integer
      required:
      - auditLogs
      - pageIndex
      - pageSize
      - totalCount
      type: object
    AuditLog:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        apiKey:
          createdAt: 2000-01-23T04:56:07.000+00:00
          lastUsedAt: 2000-01-23T04:56:07.000+00:00
          name: name
          vaults:
          - name: name
            description: description
            category: category
            uniqueId: uniqueId
            updatedAt: 2000-01-23T04:56:07.000+00:00
          - name: name
            description: description
            category: category
            uniqueId: uniqueId
            updatedAt: 2000-01-23T04:56:07.000+00:00
          id: 0
          isActive: true
          expiresAt: 2000-01-23T04:56:07.000+00:00
          updatedAt: 2000-01-23T04:56:07.000+00:00
        ipAddress: ipAddress
        action: read_vault
        userAgent: userAgent
        source: web
        vault:
          name: name
          description: description
          category: category
          uniqueId: uniqueId
          updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        createdAt:
          description: When the action occurred
          format: date-time
          type: string
        vault:
          $ref: "#/components/schemas/VaultLite"
        apiKey:
          $ref: "#/components/schemas/VaultAPIKey"
        action:
          description: Type of action performed
          enum:
          - read_vault
          - update_vault
          - delete_vault
          - create_vault
          - login_user
          - register_user
          - logout_user
          - create_api_key
          - update_api_key
          - delete_api_key
          type: string
        source:
          description: Source of the request (web interface or CLI)
          enum:
          - web
          - cli
          type: string
        ipAddress:
          description: IP address from which the action was performed
          type: string
        userAgent:
          description: User agent string from the client
          type: string
      required:
      - action
      - createdAt
      - source
      - userId
      type: object
...
    VaultAPIKey:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        lastUsedAt: 2000-01-23T04:56:07.000+00:00
        name: name
        vaults:
        - name: name
          description: description
          category: category
          uniqueId: uniqueId
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - name: name
          description: description
          category: category
          uniqueId: uniqueId
          updatedAt: 2000-01-23T04:56:07.000+00:00
        id: 0
        isActive: true
        expiresAt: 2000-01-23T04:56:07.000+00:00
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Unique API key ID
          format: int64
          type: integer
        name:
          description: Human-readable name for the API key
          type: string
        vaults:
          description: Array of vaults this key can access (null/empty = all user's
            vaults)
          items:
            $ref: "#/components/schemas/VaultLite"
          type: array
        expiresAt:
          description: Optional expiration date
          format: date-time
          type: string
        lastUsedAt:
          description: When the key was last used
          format: date-time
          type: string
        isActive:
          description: Whether the key is currently active
          type: boolean
        createdAt:
          description: When the key was created
          format: date-time
          type: string
        updatedAt:
          description: When the key was last updated
          format: date-time
          type: string
      required:
      - createdAt
      - id
      - isActive
      - name
      type: object
...
    VaultsResponse:
      example:
        pageIndex: 1
        vaults:
        - name: name
          description: description
          category: category
          uniqueId: uniqueId
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - name: name
          description: description
          category: category
          uniqueId: uniqueId
          updatedAt: 2000-01-23T04:56:07.000+00:00
        pageSize: 6
        totalCount: 0
      properties:
        vaults:
          description: Page of vault records for the authenticated user
          items:
            $ref: "#/components/schemas/VaultLite"
          type: array
        totalCount:
          description: Total number of vaults available for pagination
          type: integer
        pageSize:
          description: Number of vaults returned per page
          type: integer
        pageIndex:
          description: Current page index (starting from 1)
          type: integer
      required:
      - pageIndex
      - pageSize
      - totalCount
      - vaults
      type: object
  securitySchemes:
    apiKeyAuth:
      scheme: bearer
      type: http